+++ packages/gestalt/src/Tag.js
@@ -1,5 +1,5 @@
 // @flow strict
-import type { Node } from 'react';
+import { type Node } from 'react';
 import classnames from 'classnames';
 import Box from './Box.js';
 import Flex from './Flex.js';
@@ -34,6 +34,9 @@ type Props = {|
 
 /**
  * https://gestalt.pinterest.systems/Tag
+ *
+ * ![Tag light mode](https://raw.githubusercontent.com/pinterest/gestalt/master/cypress/integration/visual-test/__image_snapshots__/Tag%20%230.png)
+ * ![Tag dark mode](https://raw.githubusercontent.com/pinterest/gestalt/master/cypress/integration/visual-test/__image_snapshots__/Tag-dark%20%230.png)
  */
 export default function Tag(props: Props): Node {
   const {